+++ CHANGES.md
@@ -27,6 +27,7 @@ wavesurfer.js changelog
   - Added ability to specify time format for Regions tooltip using timeformatCallback (#1948)
 - Add `splitChannelsOptions` param and `setFilteredChannels` method to configure how channels are drawn (#1947)
 - Added checks in `minimap` plugin for `drawer` presence (#1953)
+- Add `setDisabledEventEmissions` method to optionally disable calls to event handlers for specific events (#1960)
 
 3.3.3 (16.04.2020)
 ------------------